requirements.txt
alembic==1.0.2
amdataset==0.2
atomicwrites==1.2.1
attrs==18.2.0
banal==0.4.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
chardet==3.0.4
#chromedriver-installer==0.0.6
dataset==1.1.0
et-xmlfile==1.0.1
fake-useragent==0.1.11
Faker==1.0.1
idna==2.7
jdcal==1.4
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.3.0
normality==0.6.1
numpy==1.15.3
openpyxl==2.5.9
pandas==0.23.4
pluggy==0.8.0
psutil==5.4.8
py==1.7.0
PyJWT==1.6.4
PySocks==1.6.8
pytest==3.9.3
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.7
requests==2.20.0
retrying==1.3.3
selenium==3.141.0
selenium-helper==0.1.2.2
six==1.11.0
text-unidecode==1.2
twilio==6.19.2
urllib3==1.24.1
